Move file_urlreader into separate set of files
