+++ include/file_urlreader.h
@@ -1,0 +1,24 @@
+#ifndef NEWSBOAT_FILE_URLREADER_H_
+#define NEWSBOAT_FILE_URLREADER_H_
+
+#include <string>
+
+#include "urlreader.h"
+
+namespace newsboat {
+
+class file_urlreader : public urlreader {
+public:
+	explicit file_urlreader(const std::string& file = "");
+	void write_config() override;
+	void reload() override;
+	void load_config(const std::string& file);
+	std::string get_source() override;
+
+private:
+	std::string filename;
+};
+
+}
+
+#endif /* NEWSBOAT_FILE_URLREADER_H_ */
