+++ mk/newsboat.deps
@@ -1,1 +1,1 @@
-newsboat.cpp src/cache.cpp  src/htmlrenderer.cpp src/urlreader.cpp src/logger.cpp src/view.cpp src/controller.cpp src/reloadthread.cpp src/tagsouppullparser.cpp src/downloadthread.cpp src/rss.cpp src/rss_parser.cpp src/formaction.cpp src/list_formaction.cpp src/feedlist_formaction.cpp src/itemlist_formaction.cpp src/itemview_formaction.cpp src/help_formaction.cpp src/filebrowser_formaction.cpp src/urlview_formaction.cpp src/select_formaction.cpp src/history.cpp src/filtercontainer.cpp src/listformatter.cpp src/regexmanager.cpp src/dialogs_formaction.cpp src/ttrss_api.cpp src/ttrss_urlreader.cpp src/newsblur_api.cpp src/newsblur_urlreader.cpp src/oldreader_urlreader.cpp src/oldreader_api.cpp src/feedcontainer.cpp src/feedhq_api.cpp src/feedhq_urlreader.cpp src/textformatter.cpp src/ocnews_api.cpp src/ocnews_urlreader.cpp src/remote_api.cpp src/inoreader_api.cpp src/inoreader_urlreader.cpp src/cliargsparser.cpp src/configpaths.cpp src/reloader.cpp src/reloadrangethread.cpp src/opml.cpp
+newsboat.cpp src/cache.cpp  src/htmlrenderer.cpp src/urlreader.cpp src/logger.cpp src/view.cpp src/controller.cpp src/reloadthread.cpp src/tagsouppullparser.cpp src/downloadthread.cpp src/rss.cpp src/rss_parser.cpp src/formaction.cpp src/list_formaction.cpp src/feedlist_formaction.cpp src/itemlist_formaction.cpp src/itemview_formaction.cpp src/help_formaction.cpp src/filebrowser_formaction.cpp src/urlview_formaction.cpp src/select_formaction.cpp src/history.cpp src/filtercontainer.cpp src/listformatter.cpp src/regexmanager.cpp src/dialogs_formaction.cpp src/ttrss_api.cpp src/ttrss_urlreader.cpp src/newsblur_api.cpp src/newsblur_urlreader.cpp src/oldreader_urlreader.cpp src/oldreader_api.cpp src/feedcontainer.cpp src/feedhq_api.cpp src/feedhq_urlreader.cpp src/textformatter.cpp src/ocnews_api.cpp src/ocnews_urlreader.cpp src/remote_api.cpp src/inoreader_api.cpp src/inoreader_urlreader.cpp src/cliargsparser.cpp src/configpaths.cpp src/reloader.cpp src/reloadrangethread.cpp src/opml.cpp src/file_urlreader.cpp
