+++ src/feedhq_urlreader.cpp
@@ -1,5 +1,6 @@
 #include "feedhq_api.h"
 
+#include "file_urlreader.h"
 #include "logger.h"
 
 namespace newsboat {
