+++ src/oldreader_urlreader.cpp
@@ -1,5 +1,6 @@
 #include "oldreader_api.h"
 
+#include "file_urlreader.h"
 #include "logger.h"
 
 namespace newsboat {
