+++ src/ttrss_urlreader.cpp
@@ -1,5 +1,6 @@
 #include "ttrss_api.h"
 
+#include "file_urlreader.h"
 #include "logger.h"
 
 namespace newsboat {
